Start separating default/saved templates
